+++ app/src/trophy.h
@@ -22,5 +22,5 @@ int trophy_prepare(const char *trophyid, char game_titleid[16]);
 void trophy_finish();
 
 int trophy_list(trophy_detail_t **details, int only_unlockable);
-int trophy_unlock(trophy_detail_t *details, int index, long int *platid, uint64_t cust_tick);
+int trophy_unlock(trophy_detail_t *details, int index, int *platid, uint64_t cust_tick);
 int trophy_unlock_all(int rand_days, uint64_t cust_tick);
